рефакторит вывод рейтинга
